Changes prices to ints

+++ src/main/kotlin/com/pts62/common/europe/ISubInvoice.kt
@@ -1,11 +1,9 @@
 package com.pts62.common.europe
 
-import java.math.BigDecimal
-
 interface ISubInvoice {
     val InvoiceNumber: String
     val Country: String
     val PaymentStatus: Boolean
     val InvoiceDate: String
-    val Price: BigDecimal
+    val Price: Int
 }
